requirements.txt
asgiref==3.2.3
certifi==2019.11.28
Django==3.0
django-crispy-forms==1.8.1
django-grappelli==2.13.2
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
wincertstore==0.2